# solution/backend/pkg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "PROD Backend 2025 Advertising Platform API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/time/advance": {
            "get": {
                "description": "Возвращает текущую дату",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Время"
                ],
                "summary": "Получить текущую дату",
                "responses": {
                    "200": {
                        "description": "Дата успешно получена",
                        "schema": {
                            "$ref": "#/definitions/model.TimeSetRequest"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Устанавливает текущую дату в кэш, только если она не меньше текущей даты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Время"
                ],
                "summary": "Установить текущую дату",
                "parameters": [
                    {
                        "description": "Запрос на установку текущей даты",
                        "name": "time",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.TimeSetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Дата успешно установлена",
                        "schema": {
                            "$ref": "#/definitions/model.TimeSetRequest"
                        }
                    },
                    "400": {
                        "description": "Неверная дата",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.ErrorResponse": {
            "type": "object",
            "required": [
                "message",
                "status"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "model.TimeSetRequest": {
            "type": "object",
            "required": [
                "current_date"
            ],
            "properties": {
                "current_date": {
                    "type": "integer",
                    "minimum": 0
                }
            }
        }
    }
}